+++ packages/coin-support-sia/src/operations/broadcastTransaction/index.ts
@@ -94,6 +94,7 @@ export const broadcastTransaction = async (
       parentAssetId: account.parentAssetId,
       familyId: account.familyId,
       parentAccountId: account.parentAccountId,
+      remarks: [transaction.userInputs.outputs[0].remarks ?? ''],
     };
 
     const [addedTxn] = await insertOrUpdateTransactions(db, [