+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cypherock/cysync-desktop
 
+## 2.4.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-desktop-ui@0.2.15
+
+## 2.4.6-pre.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-desktop-ui@0.2.15-pre.0
+
 ## 2.4.5
 
 ### Patch Changes